test: improve config instance
